requirements.txt
cycler==0.11.0
fonttools==4.31.2
kiwisolver==1.4.0
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
git+ssh://git@github.com/assured-autonomy/pyTWTL-python3.git@master#egg=pyTWTL
PyYAML==6.0
six==1.16.0
tqdm==4.63.0
pygraphviz==1.9
